refactor(layout.DirectedGraph): separate module
