+++ packages/joint-core/src/layout/index.mjs
@@ -1,5 +1,4 @@
 import * as Port from './ports/port.mjs';
 import * as PortLabel from './ports/portLabel.mjs';
 
-export * from './DirectedGraph/DirectedGraph.mjs';
 export { Port, PortLabel };
