+++ packages/joint-layout-directed-graph/.gitignore
@@ -1,0 +1,3 @@
+build/
+dist/
+node_modules/
