+++ packages/joint-layout-directed-graph/README.md
@@ -1,0 +1,11 @@
+# JointJS General Shapes
+
+A collection of general ECMAScript / TypeScript shapes for *[JointJS](https://www.jointjs.com/)*.
+
+This library fully depends on [JointJS](https://github.com/clientio/joint) (*>=3.5*), so please read its `README.md` before using this library.
+
+## License
+
+[Mozilla Public License 2.0](https://www.mozilla.org/en-US/MPL/2.0/)
+
+Copyright © 2013-2023 client IO
