+++ packages/joint-layout-directed-graph/SECURITY.md
@@ -1,0 +1,12 @@
+# Security Policy
+
+## Supported Versions
+
+The [latest released version](https://github.com/clientIO/joint/releases) of JointJS is supported.
+
+## Reporting a Vulnerability
+
+Please email security@client.io, and we will respond as quickly as possible.
+
+If the vulnerability is considered valid and accepted, a patch will be made for the latest JointJS version.
+If the vulnerability is deemed invalid, no further action is required.
