+++ packages/joint-layout-directed-graph/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfig.common.json",
+  "compilerOptions": {
+    "module": "commonjs",
+    "outDir": "./build"
+  }
+}
